+++ packages/common/CHANGELOG.md
@@ -17,6 +17,7 @@
   - `ElectionInfo` changed to `ElectionInfoV0 | ElectionInfoV1`
 - Renamed type `MicroGtuPerEuroUpdate` to `MicroCCDPerEuroUpdate`
 - Renamed enum member `UpdateType.MicroGtuPerEuro` to `UpdateType.MicroCCDPerEuro`, along with it's corresponding string value from `"microGtuPerEuro"` to `"microCCDPerEuro"`.
+- Changed type of `protocolVersion` on `BlockInfo`, `RewardStatus`, and `ConsensusStatus` from `bigint` to `ProtocolVersion` to avoid confusion when translating from the GRPC enum type.
 
 ### Fixed